Criação de métodos em Persistencia... Tela Login... Teste de Integração

+++ Dain/src/br/ufs/dain/dominio/Bolsista.java
@@ -17,7 +17,7 @@ public class Bolsista extends Aluno {
 		return tipoAtividade;
 	}
 
-	public void setTipoAtividade(short tipoAtividade) {
+	public void setTipoAtividade(int tipoAtividade) {
 		this.tipoAtividade = tipoAtividade;
 	}
 }
